Refactoring for specs directory
