+++ lib/cloudevent.js
@@ -1,5 +1,5 @@
-var Spec_0_1          = require('./spec_0_1.js');
-var Spec_0_2          = require('./spec_0_2.js');
+var Spec_0_1          = require('./specs/spec_0_1.js');
+var Spec_0_2          = require('./specs/spec_0_2.js');
 var JSONFormatter_0_1 = require('./format/json_0_1.js');
 
 function Cloudevent(_spec, _formatter){
